Tesla

Description

Tesla is intended to provide access to all known Tesla APIs as documented at https://www.teslaapi.io/

Documentation

authenticate :: AuthInfo -> IO AuthResponse

Authenticate to the Tesla service.

refreshAuth :: AuthInfo -> AuthResponse -> IO AuthResponse

Refresh authentication credentials using a refresh token.

data AuthResponse

An Authentication response.

Constructors

AuthResponse

Fields

_access_token :: String

_expires_in :: Int

_refresh_token :: String

data Product

Tesla Product Types.

Constructors

ProductVehicle

Fields

_vehicleName :: Text

_vehicleID :: VehicleID

ProductEnergy

Fields

_energyID :: EnergyID

ProductPowerWall

vehicleName :: Traversal' Product Text

vehicleID :: Traversal' Product VehicleID

energyID :: Traversal' Product EnergyID

_ProductVehicle :: Prism' Product (Text, VehicleID)

_ProductEnergy :: Prism' Product EnergyID

_ProductPowerWall :: Prism' Product ()

type VehicleID = Text

A VehicleID.

vehicles :: [Product] -> Map Text Text

Get a mapping of vehicle name to vehicle ID.

products :: MonadIO m => AuthInfo -> m [Product]

Get all products associated with this account.

type EnergyID = Integer

An energy site ID.

energyIDs :: [Product] -> [EnergyID]

Get a list of Solar ID installations.

fromToken :: String -> AuthInfo

Get an AuthInfo instance from a bearer token.

authOpts :: AuthInfo -> Options

Get a set of wreq options from an AuthInfo.

baseURL :: String

decodeProducts :: Value -> [Product]