Tesla.Car

Description

Access of car-specific APIs.

Car Monad and related types.

data Car m a

Car Monad for accessing car-specific things.

runCar :: MonadIO m => IO AuthInfo -> VehicleID -> Car m a -> m a

Run a Car Monad with the given Vehicle ID

runNamedCar :: MonadIO m => Text -> IO AuthInfo -> Car m a -> m a

Run a Car Monad by looking up a car by name.

type VehicleID = Text

A VehicleID.

Requests

vehicleData :: MonadIO m => Car m VehicleData

Fetch the VehicleData.

nearbyChargers :: MonadIO m => Car m [Charger]

Get the nearby chargers.

vehicleStatus :: MonadIO m => Car m VehicleState

vehicleStatus returns the current status of the current vehicle.

isAwake :: MonadIO m => Car m Bool

isAwake returns true if the current vehicle is awake and online.

Convenience functions for examining VehicleData

type VehicleData = ByteString

Giant blob of VehicleData describing all known state of the vehicle.

This is not separated into discrete fields because that's easy enough to do with Aeson and Lens when you need it but some convenience methods for common accesses are available in this module.

isUserPresent :: VehicleData -> Bool

True if a user is present in the vehicle.

isCharging :: VehicleData -> Bool

True of the vehicle is currently charging.

teslaTS :: VehicleData -> UTCTime

Get the timestamp from this VehicleData or error if there isn't one.

maybeTeslaTS :: VehicleData -> Maybe UTCTime

Get the timestamp from this VehicleData if present.

data Door

The various doors.

Constructors

DriverFront

DriverRear

PassengerFront

PassengerRear

FrontTrunk

RearTrunk

data OpenState a

Constructors

Closed a

Open a

_Open :: forall a. Prism' (OpenState a) a

_Closed :: forall a. Prism' (OpenState a) a

doors :: VehicleData -> Maybe [OpenState Door]

Return a list of doors and their OpenState.

openDoors :: VehicleData -> [Door]

Return a list of open doors.

Charger Info

data Location

Location, Location, Location.

Constructors

Location

Fields

_lat :: Double

_lon :: Double

data DestinationCharger

A destination charger (provided by nearbyChargers).

Constructors

DestinationCharger

Fields

_location :: Location

_name :: Text

_distance_miles :: Double

data Supercharger

A supercharger (provided by nearbyChargers).

Constructors

Supercharger

Fields

_location :: Location

_name :: Text

_distance_miles :: Double

_available_stalls :: Int

_total_stalls :: Int

_site_closed :: Bool

data Charger

Eitehr a Supercharger or Destination charger.

Constructors

SC Supercharger

DC DestinationCharger

superchargers :: [Charger] -> [Supercharger]

Return only the superchargers from a Charger list.

destinationChargers :: [Charger] -> [DestinationCharger]

Return only the destination chargers from a Charger list.

Lenses/Prisms

lat :: Lens' Location Double

lon :: Lens' Location Double

_SC :: Prism' Charger Supercharger

_DC :: Prism' Charger DestinationCharger

vdata :: Prism' VehicleData Value

Prism for viewing VehicleData as an Aeson Value.

name :: HasName s a => Lens' s a

location :: HasLocation s a => Lens' s a

distance_miles :: HasDistance_miles s a => Lens' s a

available_stalls :: HasAvailable_stalls s a => Lens' s a

total_stalls :: HasTotal_stalls s a => Lens' s a

site_closed :: HasSite_closed s a => Lens' s a

Probably uninteresting internals

vehicleURL :: VehicleID -> String -> String

Get the URL for a named endpoint for a given vehicle.

currentVehicleID :: MonadReader CarEnv m => m VehicleID

Get the current vehicle ID from the Car Monad.

Orphan instances

(Monad m, MonadIO m, MonadReader CarEnv m) => HasTeslaAuth m

Instance details

Methods

teslaAuth :: m AuthInfo