Tesla.Car.Command

Description

Executing commands within the Car Monad.

Documentation

runCmd :: (MonadIO m, Postable p) => String -> p -> Car m CommandResponse

Run a command with a payload.

runCmd' :: MonadIO m => String -> Car m CommandResponse

Run command without a payload

type CommandResponse = Either Text ()

A CommandResponse wraps an Either such that Left represents a failure message and Right suggests the command was successful.

data Car m a

Car Monad for accessing car-specific things.

TH support for generating commands.

mkCommand :: String -> String -> Q [Dec]

Build a simple named command car that posts to the given named endpoint.

mkCommands :: [String] -> Q [Dec]

Build a bunch of commands from a list of named endpoints, defining functions by removing the common prefix.

mkNamedCommands :: [(String, String)] -> Q [Dec]

Make commands with given names.

Orphan instances

FormValue Bool

Instance details

Methods

renderFormValue :: Bool -> ByteString