Tesla.Car.Command.Climate

Documentation

hvacOn :: forall {m}. MonadIO m => Car m CommandResponse

hvacOff :: forall {m}. MonadIO m => Car m CommandResponse

data ClimateKeeper

Constructors

ClimateKeeperOff

ClimateKeeperDefault

DogMode

CampMode

climateKeeper :: MonadIO m => ClimateKeeper -> Car m CommandResponse

heatSeat :: MonadIO m => Seat -> Int -> Car m CommandResponse

Set heating levels for various seats.

coolSeat :: MonadIO m => Seat -> Int -> Car m CommandResponse

Set cooling levels for various seats.

data Seat

Constructors

DriverSeat

PassengerSeat

RearLeftSeat

RearCenterSeat

RearRightSeat

setTemps :: MonadIO m => (Double, Double) -> Car m CommandResponse

Set the main HVAC temperatures.

wheelHeater :: MonadIO m => Bool -> Car m CommandResponse

Turn on the steering wheel heater

wheelHeaterOff :: MonadIO m => Car m CommandResponse

wheelHeaterOn :: MonadIO m => Car m CommandResponse

maxDefrost :: MonadIO m => Bool -> Car m CommandResponse

wakeUp :: forall {m}. MonadIO m => Car m CommandResponse

bioweaponMode :: MonadIO m => Bool -> Car m CommandResponse

Turn on or off bioweapon defense mode.

If HVAC is off, turning on bioweapon defense mode will also turn on HVAC.

data Sometimes

When configuring scheduled departure, preconditioning and off-peak charging both have weekday only options.

Constructors

Never

Always

WeekdaysOnly

data OffPeakConfig

Configuration for off-peak charging for a schedule departure.

Constructors

OffPeakConfig

Fields

_offPeakEnabled :: Sometimes

_offPeakEndTime :: Time

type Preconditioning = Sometimes

Type alias to make scheduleDeparture more readable.

scheduledDepartureOff :: MonadIO m => Car m CommandResponse

scheduleDeparture :: MonadIO m => Time -> Preconditioning -> Maybe OffPeakConfig -> Car m CommandResponse

Schedule a departure.

For this to do anything useful, you need to specify at least one of Preconditioning and/or OffPeakConfig.