Tesla.Energy

Description

Access of energy-specific APIs.

Documentation

runEnergy :: MonadIO m => IO AuthInfo -> EnergyID -> Energy m a -> m a

Run a Energy Monad with the given Vehicle ID

siteData :: (FromJSON j, MonadIO m) => Energy m j

Fetch the "live_status" describing the current active state of an energy site.

siteConfig :: (FromJSON j, MonadIO m) => Energy m j

Fetch the "site_info" describing the basic configuration of an energy site.

data Energy m a

Energy Monad for accessing energy-specific things.

Orphan instances

(Monad m, MonadIO m, MonadReader EnergyEnv m) => HasTeslaAuth m

Instance details

Methods

teslaAuth :: m AuthInfo