Tesla

Description

Tesla is intended to provide access to all known Tesla APIs as documented at https://www.teslaapi.io/

Documentation

authenticate :: AuthInfo -> IO AuthResponse

Authenticate to the Tesla service.

refreshAuth :: AuthInfo -> AuthResponse -> IO AuthResponse

Refresh authentication credentials using a refresh token.

data AuthResponse

An Authentication response.

Constructors

AuthResponse

Fields

_access_token :: String

_expires_in :: Int

_refresh_token :: String

data Product

Tesla Product Types.

Constructors

ProductVehicle

Fields

_vehicleName :: Text

_vehicleID :: VehicleID

_vehicleState :: VehicleState

ProductEnergy

Fields

_energyID :: EnergyID

ProductPowerwall

Fields

_pwID :: EnergyID

_pwBatteryPower :: Double

_pwEnergyLeft :: Double

_pwCharged :: Double

_pwName :: Text

_pwTotal :: Double

vehicleName :: Traversal' Product Text

vehicleID :: Traversal' Product VehicleID

vehicleState :: Traversal' Product VehicleState

energyID :: Traversal' Product EnergyID

_ProductVehicle :: Prism' Product (Text, VehicleID, VehicleState)

_ProductEnergy :: Prism' Product EnergyID

_ProductPowerwall :: Prism' Product (EnergyID, Double, Double, Double, Text, Double)

pwBatteryPower :: Traversal' Product Double

pwCharged :: Traversal' Product Double

pwEnergyLeft :: Traversal' Product Double

pwID :: Traversal' Product EnergyID

pwName :: Traversal' Product Text

pwTotal :: Traversal' Product Double

type VehicleID = Text

A VehicleID.

vehicles :: [Product] -> Map Text Text

Get a mapping of vehicle name to vehicle ID.

products :: MonadIO m => AuthInfo -> m [Product]

Get all products associated with this account.

productsRaw :: (FromJSON j, MonadIO m) => AuthInfo -> m j

productsRaw retrieves the complete response for products

data VehicleState

Possible states a vehicle may be in.

Constructors

VOnline

VOffline

VAsleep

VWaking

VUnknown

vsFromString :: Text -> VehicleState

type EnergyID = Integer

An energy site ID.

energyIDs :: [Product] -> [EnergyID]

Get a list of Solar ID installations.

fromToken :: String -> AuthInfo

Get an AuthInfo instance from a bearer token.

authOpts :: AuthInfo -> Options

Get a set of wreq options from an AuthInfo.

baseURL :: String

decodeProducts :: Value -> [Product]

Decode a products response into a list of products.